+++ src/Cartalyst/Sentry/SentryServiceProvider.php
@@ -130,9 +130,9 @@ class SentryServiceProvider extends ServiceProvider {
 	{
 		$this->app['sentry.group'] = $this->app->share(function($app)
 		{
-			$class = $app['config']['cartalyst/sentry::sentry.groups.model'];
+			$model = $app['config']['cartalyst/sentry::sentry.groups.model'];
 
-			return new GroupProvider($class);
+			return new GroupProvider($model);
 		});
 	}
 
@@ -145,9 +145,24 @@ class SentryServiceProvider extends ServiceProvider {
 	{
 		$this->app['sentry.user'] = $this->app->share(function($app)
 		{
-			$class = $app['config']['cartalyst/sentry::sentry.users.model'];
+			$model = $app['config']['cartalyst/sentry::sentry.users.model'];
 
-			return new UserProvider($app['sentry.hasher'], $class);
+			// We will never be accessing a user in Sentry without accessing
+			// the user provider first. So, we can lazily setup our user
+			// model's login attribute here. If you are manually using the
+			// attribute outside of Sentry, you will need to ensure you are
+			// overriding at runtime.
+			if (method_exists($model, 'setLoginAttribute'))
+			{
+				$loginAttribute = $app['config']['cartalyst/sentry::sentry.users.login_attribute'];
+
+				forward_static_call_array(
+					array($model, 'setLoginAttribute'),
+					array($loginAttribute)
+				);
+			}
+
+			return new UserProvider($app['sentry.hasher'], $model);
 		});
 	}
 
@@ -160,9 +175,9 @@ class SentryServiceProvider extends ServiceProvider {
 	{
 		$this->app['sentry.throttle'] = $this->app->share(function($app)
 		{
-			$class = $app['config']['cartalyst/sentry::sentry.throttling.model'];
+			$model = $app['config']['cartalyst/sentry::sentry.throttling.model'];
 
-			$throttleProvider = new ThrottleProvider($app['sentry.user'], $class);
+			$throttleProvider = new ThrottleProvider($app['sentry.user'], $model);
 
 			if ($app['config']['cartalyst/sentry::sentry.throttling.enabled'] === false)
 			{